% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet.R
\name{Calculate_multiplet_rate}
\alias{Calculate_multiplet_rate}
\title{Calculate_multiplet_rate for 10xGenomics Single cell 3' kit}
\usage{
Calculate_multiplet_rate(n_cell, kit = "v4")
}
\arguments{
\item{n_cell}{total cells}

\item{kit}{kit version}
}
\value{
rate
}
\description{
Calculate_multiplet_rate for 10xGenomics Single cell 3' kit
}
