% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropletqc.R
\name{EstimateEmptyDrops}
\alias{EstimateEmptyDrops}
\title{DropletQC empty drops}
\usage{
EstimateEmptyDrops(
  fnf = NULL,
  meta = NULL,
  nf_rescue = 0.05,
  umi_rescue = 1000,
  include_plot = FALSE,
  col_ct = "",
  outdir = "."
)
}
\arguments{
\item{fnf}{dropletqc data file}

\item{meta}{metadata data-frame}

\item{outdir}{out dir}
}
\description{
DropletQC empty drops
}
