% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropletqc.R
\name{RunDelEmptyDrop}
\alias{RunDelEmptyDrop}
\title{DropletQC - define and remove empty-drop}
\usage{
RunDelEmptyDrop(dir = NULL, obj = NULL, outdir = ".")
}
\arguments{
\item{dir}{cellranger out dir}

\item{obj}{seurat object}
}
\description{
DropletQC - define and remove empty-drop
}
