% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropletqc.R
\name{RunEmptyDrop}
\alias{RunEmptyDrop}
\title{DropletQC empty drops}
\usage{
RunEmptyDrop(
  nf = NULL,
  meta = NULL,
  nf_rescue = 0.05,
  umi_rescue = 1000,
  include_plot = FALSE,
  col_ct = "",
  outdir = "."
)
}
\arguments{
\item{outdir}{out dir}

\item{fnf}{dropletqc data}

\item{fmeta}{metadata}
}
\description{
DropletQC empty drops
}
