% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropletqc.R
\name{RunFilterEmptyDrops}
\alias{RunFilterEmptyDrops}
\title{Filter empty droplet using DropletQC}
\usage{
RunFilterEmptyDrops(dir = NULL, obj = NULL, form = "cellranger", outdir = ".")
}
\arguments{
\item{dir}{cellranger out dir}

\item{obj}{seurat object}
}
\description{
Filter empty droplet using DropletQC
}
