% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdblfinder.R
\name{Run_scDblFinder_ATAC}
\alias{Run_scDblFinder_ATAC}
\title{scDblFinder for ATAC}
\usage{
Run_scDblFinder_ATAC(obj = NULL, fragfile = NULL, repeats = NULL)
}
\arguments{
\item{obj}{seurat object}

\item{fragfile}{fragments.tsv.gz file}

\item{repeats}{blacklist GRanges form}
}
\description{
scDblFinder for ATAC
}
