% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet.R
\name{RunDoubletFinder}
\alias{RunDoubletFinder}
\title{DoubletFinder}
\usage{
RunDoubletFinder(obj = NULL, singlet = TRUE, outdir = ".")
}
\arguments{
\item{obj}{object}

\item{singlet}{TRUE means remove doublet, and keep only singlet}
}
\value{
object
}
\description{
DoubletFinder
}
