% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet.R
\name{RunDoubletFinder2}
\alias{RunDoubletFinder2}
\title{DoubletFinder 2}
\usage{
RunDoubletFinder2(
  obj = NULL,
  singlet = TRUE,
  rate = 0.076,
  max_dim = 10,
  outdir = "."
)
}
\arguments{
\item{obj}{object}

\item{singlet}{TRUE means remove doublet, and keep only singlet}

\item{rate}{set rate assume total cells=10000}
}
\value{
object
}
\description{
DoubletFinder 2
}
